Revert "Store this size in a fixed sized, as potentially this could overflow" This is not the bug I was looking for, and printf is annoying, so let's postpone this one for tomorrow
This reverts commit a7eafe040faedcc0987b01b878672faec769d3f7.

+++ src/smpi/private.h
@@ -412,7 +412,7 @@ extern XBT_PRIVATE double smpi_cpu_threshold;
 extern XBT_PRIVATE double smpi_host_speed;
 extern XBT_PRIVATE bool smpi_privatize_global_variables;
 extern XBT_PRIVATE char* smpi_start_data_exe; //start of the data+bss segment of the executable
-extern XBT_PRIVATE uint64_t smpi_size_data_exe; //size of the data+bss segment of the executable
+extern XBT_PRIVATE int smpi_size_data_exe; //size of the data+bss segment of the executable
 
 XBT_PRIVATE void smpi_switch_data_segment(int dest);
 XBT_PRIVATE void smpi_really_switch_data_segment(int dest);
